# Assignment 12/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SentinelPay Fraud Detection API",
    "description": "Real-time fraud detection and prevention REST API. Implements the Transaction ingestion pipeline (FR-01\u2013FR-08), Fraud Case management (FR-09\u2013FR-10), and ML Model lifecycle (FR-13\u2013FR-14) from the SentinelPay System Requirements Document.\n\n**Author:** Teboho Mokoni  \n**Assignment:** 12 \u2013 Service Layer and REST API  \n**Stack:** Python 3.12 \u00b7 FastAPI \u00b7 In-memory repositories (A11)",
    "contact": {
      "name": "Teboho Mokoni",
      "url": "https://github.com/Teboho66/SentinelPay"
    },
    "license": {
      "name": "Academic project \u2014 CPUT Postgraduate Diploma"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/fraud-cases": {
      "post": {
        "tags": [
          "Fraud Cases"
        ],
        "summary": "Create a fraud investigation case",
        "description": "FR-09: Creates a fraud case for a HARD_BLOCK transaction with HIGH or CRITICAL risk tier. Priority is automatically derived: P1 (\u22650.90), P2 (\u22650.75), P3 (<0.75). Enforces one case per transaction_id.",
        "operationId": "create_fraud_case_api_fraud_cases_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFraudCaseRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Case created and added to analyst queue",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FraudCaseResponse"
                }
              }
            }
          },
          "409": {
            "description": "Case already exists for this transaction",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Risk tier must be HIGH or CRITICAL (FR-09)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Fraud Cases"
        ],
        "summary": "Get all fraud cases",
        "description": "Returns all fraud cases. Filter by status or priority using query parameters.",
        "operationId": "get_all_cases_api_fraud_cases_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Status"
            },
            "example": "OPEN"
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Priority"
            },
            "example": "P1"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FraudCaseResponse"
                  },
                  "title": "Response Get All Cases Api Fraud Cases Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/fraud-cases/{case_id}": {
      "get": {
        "tags": [
          "Fraud Cases"
        ],
        "summary": "Get a fraud case by ID",
        "operationId": "get_case_api_fraud_cases__case_id__get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FraudCaseResponse"
                }
              }
            }
          },
          "404": {
            "description": "Case not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Fraud Cases"
        ],
        "summary": "Delete a fraud case",
        "description": "CONFIRMED cases cannot be deleted \u2014 they feed the retraining pipeline.",
        "operationId": "delete_case_api_fraud_cases__case_id__delete",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Case not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "CONFIRMED cases cannot be deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/fraud-cases/{case_id}/assign": {
      "patch": {
        "tags": [
          "Fraud Cases"
        ],
        "summary": "Assign a case to a fraud analyst",
        "description": "FR-10: Assigns an OPEN case to a named analyst. Moves case status from OPEN \u2192 IN_REVIEW.",
        "operationId": "assign_to_analyst_api_fraud_cases__case_id__assign_patch",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignAnalystRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Case assigned to analyst",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FraudCaseResponse"
                }
              }
            }
          },
          "404": {
            "description": "Case not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Case is not in OPEN state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/fraud-cases/{case_id}/resolve": {
      "patch": {
        "tags": [
          "Fraud Cases"
        ],
        "summary": "Resolve a fraud case",
        "description": "FR-10: Resolves a case as CONFIRMED or DISMISSED. CONFIRMED publishes a fraud label to the MLOps retraining pipeline (BR-FC3). DISMISSED requires a non-empty analyst note (BR-FC2).",
        "operationId": "resolve_case_api_fraud_cases__case_id__resolve_patch",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveCaseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Case resolved; fraud_label_payload present if CONFIRMED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveCaseResponse"
                }
              }
            }
          },
          "404": {
            "description": "Case not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Case must be IN_REVIEW to resolve",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "DISMISSED requires a note (BR-FC2)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssignAnalystRequest": {
        "properties": {
          "analyst_id": {
            "type": "string",
            "title": "Analyst Id",
            "description": "Analyst employee ID",
            "example": "j.mokoena"
          }
        },
        "type": "object",
        "required": [
          "analyst_id"
        ],
        "title": "AssignAnalystRequest",
        "description": "PATCH /api/fraud-cases/{id}/assign \u2014 assign case to a fraud analyst."
      },
      "CreateFraudCaseRequest": {
        "properties": {
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id",
            "example": "TXN-2026-001"
          },
          "account_id_token": {
            "type": "string",
            "title": "Account Id Token",
            "example": "acc_token_abc123"
          },
          "fraud_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Fraud Score",
            "example": 0.93
          },
          "risk_tier": {
            "type": "string",
            "title": "Risk Tier",
            "description": "Must be HIGH or CRITICAL (FR-09)",
            "example": "CRITICAL"
          },
          "shap_report_ref": {
            "type": "string",
            "title": "Shap Report Ref",
            "default": "",
            "example": "s3://shap-reports/TXN-2026-001.json"
          }
        },
        "type": "object",
        "required": [
          "transaction_id",
          "account_id_token",
          "fraud_score",
          "risk_tier"
        ],
        "title": "CreateFraudCaseRequest",
        "description": "POST /api/fraud-cases \u2014 create a fraud investigation case."
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "status_code": {
            "type": "integer",
            "title": "Status Code"
          }
        },
        "type": "object",
        "required": [
          "error",
          "detail",
          "status_code"
        ],
        "title": "ErrorResponse"
      },
...
      "FraudCaseResponse": {
        "properties": {
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "account_id_token": {
            "type": "string",
            "title": "Account Id Token"
          },
          "fraud_score": {
            "type": "number",
            "title": "Fraud Score"
          },
          "risk_tier": {
            "type": "string",
            "title": "Risk Tier"
          },
          "priority": {
            "type": "string",
            "title": "Priority"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "assigned_analyst_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Analyst Id"
          },
          "shap_report_ref": {
            "type": "string",
            "title": "Shap Report Ref"
          },
          "analyst_note": {
            "type": "string",
            "title": "Analyst Note"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "sla_breach_at": {
            "type": "string",
            "format": "date-time",
            "title": "Sla Breach At"
          },
          "is_breaching_sla": {
            "type": "boolean",
            "title": "Is Breaching Sla"
          }
        },
        "type": "object",
        "required": [
          "case_id",
          "transaction_id",
          "account_id_token",
          "fraud_score",
          "risk_tier",
          "priority",
          "status",
          "assigned_analyst_id",
          "shap_report_ref",
          "analyst_note",
          "created_at",
          "resolved_at",
          "sla_breach_at",
          "is_breaching_sla"
        ],
        "title": "FraudCaseResponse",
        "description": "Response schema for FraudCase."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResolveCaseRequest": {
        "properties": {
          "resolution": {
            "type": "string",
            "title": "Resolution",
            "description": "CONFIRMED or DISMISSED",
            "example": "CONFIRMED"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": "",
            "example": "Velocity fraud pattern confirmed via geo analysis."
          }
        },
        "type": "object",
        "required": [
          "resolution"
        ],
        "title": "ResolveCaseRequest",
        "description": "PATCH /api/fraud-cases/{id}/resolve \u2014 resolve a case."
      },
      "ResolveCaseResponse": {
        "properties": {
          "case": {
            "$ref": "#/components/schemas/FraudCaseResponse"
          },
          "fraud_label_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fraud Label Payload"
          }
        },
        "type": "object",
        "required": [
          "case"
        ],
        "title": "ResolveCaseResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}